+++ packages/web-app-files/src/mixins/spaces/actions/setReadme.js
@@ -3,8 +3,14 @@ import { mapMutations, mapState, mapActions } from 'vuex'
 import { bus } from 'web-pkg/src/instance'
 
 export default {
+  inject: {
+    currentSpace: {
+      default: null
+    }
+  },
   computed: {
     ...mapState('Files', ['currentFolder']),
+    ...mapState(['user']),
     $_setSpaceReadme_items() {
       return [
         {
@@ -21,13 +27,24 @@ export default {
             if (!resources[0].mimeType?.startsWith('text/')) {
               return false
             }
-            return isLocationSpacesActive(this.$router, 'files-spaces-project')
+            if (!isLocationSpacesActive(this.$router, 'files-spaces-project')) {
+              return false
+            }
+
+            if (!this.space) {
+              return false
+            }
+
+            return this.space.canEditReadme({ user: this.user })
           },
           canBeDefault: false,
           componentType: 'oc-button',
           class: 'oc-files-actions-set-space-readme-trigger'
         }
       ]
+    },
+    space() {
+      return this.currentSpace?.value
     }
   },
   methods: {