+++ packages/web-app-files/tests/unit/views/spaces/Project.spec.js
@@ -7,10 +7,19 @@ import mockAxios from 'jest-mock-axios'
 import SpaceProject from '../../../../src/views/spaces/Project.vue'
 import Vuex from 'vuex'
 import { ShareTypes, spaceRoleManager } from '../../../../src/helpers/share'
+import { thumbnailService } from '../../../../src/services'
 
 localVue.use(GetTextPlugin, {
   translations: 'does-not-matter.json',
   silent: true
+})
+
+beforeAll(() => {
+  thumbnailService.initialize({
+    enabled: true,
+    version: '0.1',
+    supportedMimeTypes: ['image/png', 'image/jpg', 'image/jpeg', 'image/gif', 'text/plain']
+  })
 })
 
 beforeEach(mockAxios.reset)
@@ -39,6 +48,7 @@ const spaceMocks = {
   spaceWithReadmeAndImage: {
     id: 1,
     name: 'space',
+    root: { permissions: [{ roles: ['manager'], grantedTo: [{ user: { id: 1 } }] }] },
     special: [
       {
         specialFolder: { name: 'readme' },
@@ -55,6 +65,7 @@ const spaceMocks = {
   spaceWithoutReadmeAndImage: {
     id: 1,
     name: 'space',
+    root: { permissions: [{ roles: ['manager'], grantedTo: [{ user: { id: 1 } }] }] },
     special: []
   }
 }
@@ -220,6 +231,8 @@ function getMountedWrapper(spaceResources = [], spaceItem = null, imageContent =
       $client: {
         files: {
           getFileContents: jest.fn().mockImplementation(() => Promise.resolve('filecontent')),
+          // TODO: Add & reference space fixtures in line below
+          // fileInfo: jest.fn().mockImplementation(() => Promise.resolve(Files['/'][4])),
           list: jest.fn(() => spaceResources)
         }
       }
@@ -232,7 +245,11 @@ function getMountedWrapper(spaceResources = [], spaceItem = null, imageContent =
           options: {
             disablePreviews: true
           }
-        })
+        }),
+        user: () => ({
+          id: 'marie'
+        }),
+        getToken: jest.fn()
       },
       modules: {
         Files: {